% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{psychdsish-package}
\alias{psychdsish}
\alias{psychdsish-package}
\title{psychdsish}
\description{
An R package for creating an R/tidyverse project skeleton that is closely, but not entirely, compliant-ish with the [psych-DS standard](https://psych-ds.github.io/) standard for psychological data, and check whether a given project is compliant-ish with psych-DS. Hence, psych-DS-ish. The layout is loosely inspired by the psych-DS specification, but is not fully psych-DS compliant as it does not comply with psych-ds's .json requirement. Psych-ds-ish also adds extra conventions for reproducibility and common tidyverse workflows, including generating Quarto file templates, a readme, license, and a .gitignore file.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@unibe.ch} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

}
