% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{psychdsish-package}
\alias{psychdsish}
\alias{psychdsish-package}
\title{psychdsish}
\description{
An R package for creating an R/tidyverse project skeleton that is closely, but not entirely, compliant-ish with the psych-DS standard, and check whether a given project is compliant-ish with psych-DS. Hence, psych-DS(ish). The layout is loosely inspired by the [psych-DS](https://psych-ds.github.io/) specification, but is not fully psych-DS compliant. It adds extra conventions for reproducibility and common tidyverse workflows, including Quarto file templates and a .gitignore file. It does not comply with psych-ds's .json requirement.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@unibe.ch} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

}
